requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.6.3
beautifulsoup4==4.13.3
cachetools==4.2.2
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
instagram-private-api==1.6.0.0
instagrapi==2.1.3
instaloader==4.14.1
pillow==11.1.0
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
PySocks==1.7.1
python-dotenv==1.1.0
python-telegram-bot==13.7
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tornado==6.4.2
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
yt-dlp==2025.3.27
